添加商品信息显示，更新readme
