+++ README.md
@@ -8,7 +8,7 @@
 
 + `authority-server` 权限管理中心
 + `gateway-server` 网关服务
-+ `menu-server` 菜单服务
++ `goods-server` 商品服务
 + `user-server` 用户服务
 + `common-api` 通用模块
 
@@ -23,6 +23,14 @@ A(前端界面) --> B(gateway-server)
     C --> F(other-server)
     G[简陋流程]
 ```
+# 技术栈
++ nacos 注册中心、配置服务
++ gateway 网关
++ openfeign 服务调用
++ shiro 安全管理
++ redis 数据存储 
++ mysql 数据库
++ mybatisplus 
 
 *太菜了我*
