+++ authority-server-4000/src/main/java/com/hyx/authority/service/imp/GoodsServiceImpl.java
@@ -1,17 +1,19 @@
 package com.hyx.authority.service.imp;
 
-import com.hyx.authority.service.UserService;
+import com.hyx.authority.service.GoodsService;
 import com.hyx.common.entities.CommonResult;
 import org.springframework.stereotype.Service;
 
 /**
  * @author : xiaolang
- * @date ：Created in 2020/4/15 19:34
+ * @date ：Created in 2020/4/17 15:48
  */
+
 @Service
-public class UserServerImp implements UserService {
+public class GoodsServiceImpl implements GoodsService{
+
     @Override
-    public CommonResult getUserList(int current,int page) {
+    public CommonResult getGoodsList(int current, int size) {
         return new CommonResult<>(400,"请求超时","请求超时");
     }
 }
