--- common-api/src/main/java/com/hyx/common/config/Swagger2Config.java
@@ -1,45 +1,0 @@
-package com.hyx.common.config;
-
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import springfox.documentation.builders.ApiInfoBuilder;
-import springfox.documentation.builders.PathSelectors;
-import springfox.documentation.builders.RequestHandlerSelectors;
-import springfox.documentation.service.ApiInfo;
-import springfox.documentation.service.Contact;
-import springfox.documentation.spi.DocumentationType;
-import springfox.documentation.spring.web.plugins.Docket;
-import springfox.documentation.swagger2.annotations.EnableSwagger2;
-
-/**
- * @author : xiaolang
- * @date ：Created in 2020/4/16 11:31
- */
-
-@Configuration
-@EnableSwagger2
-public class Swagger2Config {
-
-    @Bean
-    public Docket createRestApi() {
-        String name = "com.hyx.authority.controller";
-        return new Docket(DocumentationType.SWAGGER_2)
-                .apiInfo(apiInfo())
-                .select()
-                .apis(RequestHandlerSelectors.basePackage(name))
-                .paths(PathSelectors.any())
-                .build();
-    }
-
-    private ApiInfo apiInfo() {
-        return new ApiInfoBuilder()
-                .title("viboot-swagger2")
-                .description("Restful-API-Doce")
-                .termsOfServiceUrl("https://www.baidu.com")
-                .contact(new Contact("小小浪", "https://127.0.0.1:8080/hello", "huangyouxin123@qq.com"))
-                .version("1.0")
-                .build();
-    }
-
-}
