+++ common-api/src/main/java/com/hyx/common/entities/SpUser.java
@@ -1,22 +1,33 @@
 package com.hyx.common.entities;
 
+import com.baomidou.mybatisplus.annotation.IdType;
+import com.baomidou.mybatisplus.annotation.TableId;
+import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
-import lombok.AllArgsConstructor;
 import lombok.Data;
-import lombok.NoArgsConstructor;
+import lombok.EqualsAndHashCode;
+import lombok.experimental.Accessors;
 
 import java.io.Serializable;
 
 /**
- * @author : xiaolang
- * @date ：Created in 2020/4/13 17:18
+ * <p>
+ * 会员表
+ * </p>
+ *
+ * @author xiaolang
+ * @since 2020-04-15
  */
-
 @Data
-@AllArgsConstructor
-@NoArgsConstructor
+@EqualsAndHashCode(callSuper = false)
+@Accessors(chain = true)
+@ApiModel(value="SpUser对象", description="会员表")
 public class SpUser implements Serializable {
+
+    private static final long serialVersionUID = 1L;
+
     @ApiModelProperty(value = "自增id")
+    @TableId(value = "user_id", type = IdType.AUTO)
     private Integer userId;
 
     @ApiModelProperty(value = "登录名")
