+++ common-api/src/main/resources/META-INF/spring.factories
@@ -1,2 +1,2 @@
-org.springframework.boot.autoconfigure.EnableAutoConfiguration=com.hyx.common.config.Swagger2Config
+#org.springframework.boot.autoconfigure.EnableAutoConfiguration=com.hyx.common.config.Swagger2Config
