+++ goods-server-4002/src/main/java/com/hyx/goods/GoodsMain.java
@@ -1,15 +1,15 @@
-package com.hyx.user;
+package com.hyx.goods;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 /**
  * @author : xiaolang
- * @date ：Created in 2020/4/15 18:12
+ * @date ：Created in 2020/4/14 19:02
  */
 @SpringBootApplication
-public class UserMain {
+public class GoodsMain {
     public static void main(String[] args) {
-        SpringApplication.run(UserMain.class,args);
+        SpringApplication.run(GoodsMain.class,args);
     }
 }
