+++ goods-server-4002/src/main/java/com/hyx/goods/service/MenuService.java
@@ -1,0 +1,13 @@
+package com.hyx.goods.service;
+
+import com.hyx.common.entities.Menu;
+
+import java.util.List;
+
+/**
+ * @author : xiaolang
+ * @date ：Created in 2020/4/14 19:05
+ */
+public interface MenuService {
+    List<Menu> getMenu();
+}
