+++ goods-server-4002/src/main/resources/application.yml
@@ -1,9 +1,9 @@
 server:
-  port: 4003
+  port: 4002
 
 spring:
   application:
-    name: user-server
+    name: goods-server
 
   #开启热部署
   devtools:
@@ -14,7 +14,7 @@ spring:
   datasource:
     type: com.alibaba.druid.pool.DruidDataSource      #当前数据源操作类型
     driver-class-name:  com.mysql.cj.jdbc.Driver     #mysql驱动包
-    url: jdbc:mysql://localhost:3306/management?useUnicode=true&serverTimezone=GMT&useSSL=false&characterEncoding=utf8
+    url: jdbc:mysql://localhost:3306/management?useUnicode=true&characterEncoding-utr-8&useSSL=false&serverTimezone=UTC
     username: root
     password: hyx2215
 
@@ -24,12 +24,11 @@ spring:
       discovery:
         server-addr: localhost:8848
 
-#配置 mybatis-plus
+#配置mybatis-plus
 mybatis-plus:
   mapper-locations:  classpath:mapper/*.xml
-  #开启 SQL 语句
   configuration:
     log-impl: org.apache.ibatis.logging.stdout.StdOutImpl
 
 moudle:
-  name: user
+  name: goods
