--- menu-server-4002/src/main/java/com/hyx/menu/MenuMain.java
@@ -1,15 +1,0 @@
-package com.hyx.menu;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-/**
- * @author : xiaolang
- * @date ：Created in 2020/4/14 19:02
- */
-@SpringBootApplication
-public class MenuMain {
-    public static void main(String[] args) {
-        SpringApplication.run(MenuMain.class,args);
-    }
-}
