--- menu-server-4002/src/main/java/com/hyx/menu/controller/MenuController.java
@@ -1,31 +1,0 @@
-package com.hyx.menu.controller;
-
-import com.alibaba.fastjson.JSONObject;
-import com.hyx.common.entities.CommonResult;
-import com.hyx.common.entities.Menu;
-import com.hyx.common.entities.MenuBase;
-import com.hyx.menu.service.MenuService;
-import io.swagger.annotations.ApiOperation;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import javax.annotation.Resource;
-import java.util.List;
-
-/**
- * @author : xiaolang
- * @date ：Created in 2020/4/14 19:03
- */
-@RestController
-public class MenuController {
-
-    @Resource
-    MenuService menuService;
-
-    @GetMapping(value = "/getMenuList")
-    @ApiOperation(value = "获取菜单列表", notes = "备注")
-    public CommonResult getMenuList(){
-        List<Menu> menus = menuService.getMenu();
-        return new CommonResult<>(200,"成功获取菜单列表",menus);
-    }
-}
