--- menu-server-4002/src/main/resources/application.yml
@@ -1,34 +1,0 @@
-server:
-  port: 4002
-
-spring:
-  application:
-    name: menu-server
-
-  #开启热部署
-  devtools:
-    restart:
-      enabled: true
-
-  #配置mysql数据库
-  datasource:
-    type: com.alibaba.druid.pool.DruidDataSource      #当前数据源操作类型
-    driver-class-name:  com.mysql.cj.jdbc.Driver     #mysql驱动包
-    url: jdbc:mysql://localhost:3306/management?useUnicode=true&characterEncoding-utr-8&useSSL=false&serverTimezone=UTC
-    username: root
-    password: hyx2215
-
-  #配置nacos
-  cloud:
-    nacos:
-      discovery:
-        server-addr: localhost:8848
-
-#配置mybatis-plus
-mybatis-plus:
-  mapper-locations:  classpath:mapper/*.xml
-#  configuration:
-#    log-impl: org.apache.ibatis.logging.stdout.StdOutImpl
-
-moudle:
-  name: authority
